Move `Model` to dedicated directory
